Fixed interrupt related deadlocks
Interrupt handlers use I2C and SPI comms with FPGA and ADV7611. If these occurred while other code
was in the middle of a transaction then the I2C/SPI controller would be left in an unexpected
state and the code could be left waiting for data to arrive that had already been flushed out

Also, the STM32 doesn't support level triggered interrupts and it's possible to miss a rising-edge.
Added a polling function to check for missed interrupts every 250ms. Without this there would
be no more interrupts from the FPGA as it would wait forever for the missed one to be cleared